chore: prevent transactional self-invocation on ProspectService::saveAllWithEmailCheck
